Add the data fixture
