+++ pygmt/tests/test_clib.py
@@ -23,15 +23,6 @@ from pygmt.exceptions import (
 from pygmt.helpers import GMTTempFile
 
 TEST_DATA_DIR = os.path.join(os.path.dirname(__file__), "data")
-POINTS_DATA = os.path.join(TEST_DATA_DIR, "points.txt")
-
-
-@pytest.fixture(scope="module", name="data")
-def fixture_data():
-    """
-    Load the point data from the test file.
-    """
-    return np.loadtxt(POINTS_DATA)
 
 
 @contextmanager
